closed #4559: Wrong behavior of multi-touch on iphone5s/ipadAir (arm64 arch)

+++ cocos/base/CCConsole.h
@@ -137,7 +137,7 @@ protected:
     std::mutex _DebugStringsMutex;
     std::vector<std::string> _DebugStrings;
 
-    int _touchId;
+    intptr_t _touchId;
 private:
     CC_DISALLOW_COPY_AND_ASSIGN(Console);
 };
